add: offeredCourse deleting logic added

+++ src/app/modules/offeredCourse/offeredCourse.service.ts
@@ -205,6 +205,18 @@ const deleteOfferedCourseFromDB = async (id: string) => {
     throw new AppError(httpStatus.NOT_FOUND, 'Offered course is not found!');
   }
 
+  const semesterRegistration = isOfferedCourseExists?.semesterRegistration;
+
+  const semesterRegistrationStatus =
+    await SemesterRegistration.findById(semesterRegistration).select('status');
+
+  if (semesterRegistrationStatus?.status !== 'UPCOMING') {
+    throw new AppError(
+      httpStatus.BAD_REQUEST,
+      `Offered course can not Delete! because the semester ${semesterRegistrationStatus}`,
+    );
+  }
+
   const result = await OfferedCourse.findByIdAndDelete(id);
   return result;
 };
